Skip checking Docker compose file changes
Not necessary since the up command does that for us already.
This will let us restart the Docker containers even though the
config hasn't changed.

Will also cause the Docker compose status message to appear
even though the containers are never restarted. For example
when a config value has changed that does not affect the
Docker compose file. We'll have to live with that for now.